spec fix

+++ Sources/Vector/mat_vec2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   mat_vec2.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lomasse <lomasse@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cbilga <cbilga@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/13 18:27:18 by lomasse           #+#    #+#             */
-/*   Updated: 2020/01/13 18:27:22 by lomasse          ###   ########.fr       */
+/*   Updated: 2020/07/26 10:52:56 by cbilga           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ t_point		cross_vec(t_point v1, t_point v2)
 
 float		dot_product(t_point v1, t_point v2)
 {
-	return ((v1.x * v2.x) + (v1.y * v2.y) + (v1.z * v2.z));
+	return (((double)v1.x * (double)v2.x) + ((double)v1.y * (double)v2.y) + ((double)v1.z * (double)v2.z));
 }
 
 float		length(t_point v1)
